separate RCSS types from Preferences
